Move constant to _app
