+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import 'styles/globals.css';
 import 'styles/markdown.css';
 import type { AppProps } from 'next/app';
-import { Layout, PLAUSIBLE_DOMAINS } from 'components/sections/Layout/Layout';
+import { Layout } from 'components/sections/Layout/Layout';
 import { appWithTranslation } from 'next-i18next';
 import '@fontsource/roboto/900.css';
 import '@fontsource/roboto/700.css';
@@ -16,7 +16,13 @@ import '@fontsource/source-sans-pro/600.css';
 import { useRouter } from 'next/router';
 import Head from 'next/head';
 import PlausibleProvider from 'next-plausible';
-import { getCountryVariant } from '../utils/locales';
+import { CountryVariant, getCountryVariant } from '../utils/locales';
+
+export const PLAUSIBLE_DOMAINS: Record<CountryVariant, string> = {
+  cs: 'movapp.cz, all.movapp.eu',
+  sk: 'sk.movapp.eu, all.movapp.eu',
+  pl: 'pl.movapp.eu, all.movapp.eu',
+};
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
   const { locales, asPath } = useRouter();
@@ -24,6 +30,7 @@ const MyApp = ({ Component, pageProps }: AppProps) => {
     <PlausibleProvider domain={PLAUSIBLE_DOMAINS[getCountryVariant()]}>
       <Head>
         <link rel="alternate" hrefLang="x-default" href={`https://www.movapp.cz${asPath}`} />
+        {/* <Script data-domain={PLAUSIBLE_DOMAINS[getCountryVariant()]} src="https://plausible.io/js/plausible.js" /> */}
         {locales?.map((locale, index) => {
           return <link key={index} rel="alternate" hrefLang={locale} href={`https://www.movapp.cz/${locale}${asPath}`} />;
         })}
